requirements.txt
streamlit
pillow
pandas
openpyxl
pytesseract
requests
opencv-python-headless
numpy